+++ sirius_gui/src/main/java/de/unijena/bioinf/ms/gui/compute/jjobs/Jobs.java
@@ -229,6 +229,7 @@ public class Jobs {
             synchronized (ACTIVE_COMPUTATIONS) { //todo this is a bit hacky but much mor efficient than listening to the job states
                 ACTIVE_COMPUTATIONS.add(this);
                 ((ShowJobsDialogAction) SiriusActions.SHOW_JOBS.getInstance()).setComputing(!ACTIVE_COMPUTATIONS.isEmpty());
+                SiriusActions.SUMMARY_WS.getInstance().setEnabled(ACTIVE_COMPUTATIONS.isEmpty());
             }
             checkForInterruption();
             compoundsToProcess.forEach(i -> i.setComputing(true));
@@ -249,6 +250,7 @@ public class Jobs {
             synchronized (ACTIVE_COMPUTATIONS) {
                 ACTIVE_COMPUTATIONS.remove(this);
                 ((ShowJobsDialogAction) SiriusActions.SHOW_JOBS.getInstance()).setComputing(!ACTIVE_COMPUTATIONS.isEmpty());
+                SiriusActions.SUMMARY_WS.getInstance().setEnabled(ACTIVE_COMPUTATIONS.isEmpty());
             }
             compoundsToProcess.forEach(i -> i.setComputing(false));
             super.cleanup();